requirements.txt
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4        
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0   
exceptiongroup==1.3.1    
fastapi==0.111.1
fastapi-cli==0.0.20      
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.2.6
langchain-core==1.2.7
langgraph==1.0.6
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.6
langgraph-sdk==0.3.3
langsmith==0.6.4
loguru==0.7.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
numpy==2.2.6
openai==1.27.0
orjson==3.11.5
ormsgpack==1.12.2
packaging==25.0
pgvector==0.4.2
psycopg2-binary==2.9.9
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
python-dotenv==1.0.0
python-multipart==0.0.21
PyYAML==6.0.3
requests==2.31.0
requests-toolbelt==1.0.0
rich==14.2.0
rich-toolkit==0.17.1
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.46
starlette==0.37.2
tenacity==9.1.2
tomli==2.4.0
tqdm==4.67.1
typer==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.13.0
uvicorn==0.23.2
watchfiles==1.1.1
websockets==16.0
win32_setctime==1.2.0
xxhash==3.6.0
zstandard==0.25.0